Move query construct and its methods to query.go
