+++ papergres.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"reflect"
 	"strconv"
 	"strings"
 	"sync"
@@ -89,14 +88,6 @@ type Schema struct {
 	Database *Database
 }
 
-// Query holds the SQL to execute and the connection string
-type Query struct {
-	SQL      string
-	Database *Database
-	Args     []interface{}
-	insert   bool
-}
-
 // Repeat holds the iteration count and params function
 // for executing a query N times.
 type Repeat struct {
@@ -104,10 +95,6 @@ type Repeat struct {
 	ParamsFn SelectParamsFn
 	N        int
 }
-
-// SelectParamsFn is a function that takes in the iteration and
-// returns the destination and args for a SQL execution.
-type SelectParamsFn func(i int) (dest interface{}, args []interface{})
 
 // Result holds the results of an executed query
 type Result struct {
@@ -147,27 +134,6 @@ func NewResult() *Result {
 		RowsAffected: RowsAffected{},
 	}
 	return result
-}
-
-func (q *Query) String() string {
-	return fmt.Sprintf(`
-	Query:
-	%s
-	%s
-	Connection: %s
-	`,
-		q.SQL, argsToString(q.Args), prettifyConnString(q.Database.ConnectionString()))
-}
-
-func argsToString(args []interface{}) string {
-	s := ""
-	if len(args) > 0 {
-		s = "Args:"
-	}
-	for i, a := range args {
-		s += fmt.Sprintf("\n\t$%v: %v", i+1, a)
-	}
-	return s
 }
 
 func (r *Result) String() string {
@@ -217,30 +183,6 @@ func Shutdown() {
 		if err := db.Close(); err != nil {
 			log.Fatalf("Error shutting down DB: %s", err.Error())
 		}
-	}
-}
-
-// Repeat will execute a query N times. The param selector function will pass in the
-// current iteration and expect back the destination obj and args for that index.
-// Make sure to use pointers to ensure the sql results fill your structs.
-// Use this when you want to run the same query for many different parameters, like getting
-// data for child entities for a collection of parents.
-// This function executes the iterations concurrently so each loop should not rely on state
-// from a previous loops execution. The function should be extremely fast and efficient with DB resources.
-// Returned error will contain all errors that occurred in any iterations.
-//
-//		params := func(i int) (dest interface{}, args []interface{}) {
-//			p := &parents[i] 						// get parent at i to derive parameters
-//			args := MakeArgs(p.Id, true, "current") // create arg list, variadic
-//			return &p.Child, args 					// &p.Child will be filled with returned data
-//		}
-//		// len(parents) is parent slice and determines how many times to execute query
-//		results, err := db.Query(sql).Repeat(len(parents), params).Exec()
-func (q *Query) Repeat(times int, pSelectorFn SelectParamsFn) *Repeat {
-	return &Repeat{
-		q,
-		pSelectorFn,
-		times,
 	}
 }
 
@@ -456,46 +398,6 @@ func (r *Repeat) Exec() ([]*Result, error) {
 	return results, mergeErrs(errs)
 }
 
-// ExecAll gets many rows and populates the given slice
-// dest should be a pointer to a slice
-func (q *Query) ExecAll(dest interface{}) *Result {
-	all := func(db *sqlx.DB, r *Result) error {
-		err := db.Select(dest, q.SQL, q.Args...)
-
-		r.RowsReturned = getLen(dest)
-
-		return err
-	}
-
-	return execDB(q, all)
-}
-
-// ExecSingle fetches a single row from the database and puts it into dest.
-// If more than 1 row is returned it takes the first one.
-// Expects at least 1 row or it will error.
-func (q *Query) ExecSingle(dest interface{}) *Result {
-	single := func(db *sqlx.DB, r *Result) error {
-		err := db.Get(dest, q.SQL, q.Args...)
-		if err == nil {
-			r.RowsReturned = 1
-		}
-		return err
-	}
-
-	return execDB(q, single)
-}
-
-// Exec runs a sql command given a connection and expects LastInsertId or RowsAffected
-// to be returned by the script. Use this for INSERTs
-func (q *Query) Exec() *Result {
-	return exec(q, false)
-}
-
-// ExecNonQuery runs the SQL and doesn't look for any results
-func (q *Query) ExecNonQuery() *Result {
-	return exec(q, true)
-}
-
 type meta struct {
 	LastInsertId PrimaryKey
 	RowsAffected int64
@@ -559,12 +461,4 @@ func mergeErrs(errs []error) error {
 		return errors.New(s)
 	}
 	return nil
-}
-
-func getLen(i interface{}) int {
-	val := reflect.Indirect(reflect.ValueOf(i))
-	if val.Kind() == reflect.Slice {
-		return val.Len()
-	}
-	return 0
 }
